Conflicts:
	devops/ansible/playbook.yml
	girder/api/v1/assetstore.py
	girder/api/v1/resource.py
	plugins/geospatial/server/geospatial.py

+++ tests/packaging/test_plugin_install.sh
@@ -2,9 +2,15 @@
 
 virtualenv_activate="${1}"
 source_path="${2}"
+virtualenv_pip="${3}"
 unset PYTHONPATH
 
 source "${virtualenv_activate}"
+
+# Mock npm executable to save time
+oldpath=$PATH
+export PATH="$(pwd)/mockpath:$PATH"
+rm -f npm_called.txt
 
 girder-install plugin "${source_path}/plugins/thumbnails"
 
@@ -13,7 +19,19 @@ if [ $? -eq 0 ] ; then
     exit 1
 fi
 
+if [ -f "npm_called.txt" ] ; then
+    echo "Npm should not have been called in failure case"
+    exit 1
+fi
+
 girder-install plugin -f "${source_path}/plugins/thumbnails" "${source_path}/plugins/oauth" || exit 1
+
+if [ ! -f "npm_called.txt" ] ; then
+    echo "Error: npm was not called during plugin install"
+    exit 1
+fi
+
+rm -f npm_called.txt
 
 dest_path=$(girder-install plugin-path)
 girder-install plugin -f "${dest_path}/thumbnails" || exit 1
@@ -23,4 +41,23 @@ if [ ! -f "${dest_path}/thumbnails/plugin.yml" ] ; then
     exit 1
 fi
 
-exit 0
+if [ ! -f "npm_called.txt" ] ; then
+    echo "Error: npm was not called during plugin install using plugin-path command"
+    exit 1
+fi
+
+rm -f npm_called.txt
+
+# Actually run the web build and make sure it works for plugins
+export PATH=$oldpath
+
+extras_path="$(girder-install web-root)/static/built/plugins/plugin_with_extras"
+echo $extras_path
+rm -fr "$extras_path"
+
+girder-install plugin -f "${source_path}/tests/packaging/plugin_with_extras" || exit 1
+
+if [ ! -f "${extras_path}/extra/data.txt" ] ; then
+    echo "Error: extra file not copied to the web root"
+    exit 1
+fi
